@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2020 ReactiveUI Association Inc. All rights reserved.
-// ReactiveUI Association Inc licenses this file to you under the MIT license.
+﻿// Copyright (c) 2019-2021 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
 using System;
@@ -11,16 +11,28 @@ using System.Runtime.InteropServices;
 
 namespace ReactiveMarbles.NuGet.Helpers
 {
-    internal static class AssemblyHelpers
+    /// <summary>
+    /// Helps get details about assemblies.
+    /// </summary>
+    public static class AssemblyHelpers
     {
         private static readonly string[] AssemblyFileExtensions =
         {
             ".winmd", ".dll", ".exe"
         };
 
+        /// <summary>
+        /// Gets the assembly file extensions set.
+        /// </summary>
         public static ISet<string> AssemblyFileExtensionsSet { get; } = new HashSet<string>(AssemblyFileExtensions, StringComparer.InvariantCultureIgnoreCase);
 
-        internal static string FindUnionMetadataFile(string name, Version version)
+        /// <summary>
+        /// Finds the union metadata file.
+        /// </summary>
+        /// <param name="name">The name.</param>
+        /// <param name="version">The version.</param>
+        /// <returns>The file if found.</returns>
+        public static string? FindUnionMetadataFile(string name, Version version)
         {
             var basePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86), "Windows Kits", "10", "UnionMetadata");
 
@@ -38,15 +50,16 @@ namespace ReactiveMarbles.NuGet.Helpers
 
             var file = Path.Combine(basePath, name + ".winmd");
 
-            if (!File.Exists(file))
-            {
-                return null;
-            }
-
-            return file;
+            return !File.Exists(file) ? null : file;
         }
 
-        internal static string FindWindowsMetadataFile(string name, Version version)
+        /// <summary>
+        /// Finds the windows metadata file.
+        /// </summary>
+        /// <param name="name">The name.</param>
+        /// <param name="version">The version.</param>
+        /// <returns>The file if found.</returns>
+        public static string? FindWindowsMetadataFile(string name, Version? version)
         {
             // This is only supported on windows at the moment.
             if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
@@ -61,6 +74,11 @@ namespace ReactiveMarbles.NuGet.Helpers
                 return FindWindowsMetadataInSystemDirectory(name);
             }
 
+            if (version is null)
+            {
+                return null;
+            }
+
             basePath = Path.Combine(basePath, FindClosestVersionDirectory(basePath, version));
 
             if (!Directory.Exists(basePath))
@@ -70,37 +88,27 @@ namespace ReactiveMarbles.NuGet.Helpers
 
             var file = Path.Combine(basePath, name + ".winmd");
 
-            if (!File.Exists(file))
-            {
-                return FindWindowsMetadataInSystemDirectory(name);
-            }
-
-            return file;
+            return !File.Exists(file) ? FindWindowsMetadataInSystemDirectory(name) : file;
         }
 
-        private static string FindWindowsMetadataInSystemDirectory(string name)
+        private static string? FindWindowsMetadataInSystemDirectory(string name)
         {
             var file = Path.Combine(Environment.SystemDirectory, "WinMetadata", name + ".winmd");
-            if (File.Exists(file))
-            {
-                return file;
-            }
-
-            return null;
+            return File.Exists(file) ? file : null;
         }
 
-        private static string FindClosestVersionDirectory(string basePath, Version version)
+        private static string? FindClosestVersionDirectory(string basePath, Version version)
         {
-            string path = null;
+            string? path = null;
             foreach (var folder in new DirectoryInfo(basePath)
                 .EnumerateDirectories()
                 .Select(d => ConvertToVersion(d.Name))
-                .Where(v => v.Item1 != null)
-                .OrderByDescending(v => v.Item1))
+                .Where(v => v.Version != null)
+                .OrderByDescending(v => v.Version))
             {
-                if (path == null || folder.Item1 >= version)
+                if (path == null || folder.Version >= version)
                 {
-                    path = folder.Item2;
+                    path = folder.Name;
                 }
             }
 
@@ -108,7 +116,7 @@ namespace ReactiveMarbles.NuGet.Helpers
         }
 
         [SuppressMessage("Design", "CA1031: Modify to catch a more specific exception type, or rethrow the exception.", Justification = "Deliberate usage.")]
-        private static (Version Version, string Name) ConvertToVersion(string name)
+        private static (Version? Version, string? Name) ConvertToVersion(string name)
         {
             string RemoveTrailingVersionInfo()
             {
