@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2020 ReactiveUI Association Inc. All rights reserved.
-// ReactiveUI Association Inc licenses this file to you under the MIT license.
+﻿// Copyright (c) 2019-2021 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
 using System;
@@ -12,7 +12,10 @@ using NuGet.Frameworks;
 
 namespace ReactiveMarbles.NuGet.Helpers
 {
-    internal static class NuGetFrameworkFolderHelper
+    /// <summary>
+    /// A folder helper to get nuget framework information.
+    /// </summary>
+    public static class NuGetFrameworkFolderHelper
     {
         /// <summary>
         /// Handles getting additional reference libraries where none exists.
@@ -21,28 +24,15 @@ namespace ReactiveMarbles.NuGet.Helpers
         /// <returns>A list of additional paths.</returns>
         public static IEnumerable<string> GetNuGetFrameworkFolders(this NuGetFramework framework)
         {
-            IEnumerable<string> folders;
-            switch (framework.Framework.ToLowerInvariant())
+            var folders = framework.Framework.ToLowerInvariant() switch
             {
-                case "monoandroid":
-                    folders = HandleAndroid(framework);
-                    break;
-                case "xamarin.ios":
-                    folders = HandleiOS();
-                    break;
-                case "xamarin.tvos":
-                    folders = HandleTVOS();
-                    break;
-                case "xamarin.watchos":
-                    folders = HandleWatchOS();
-                    break;
-                case "xamarin.mac":
-                    folders = HandleMac();
-                    break;
-                default:
-                    folders = Array.Empty<string>();
-                    break;
-            }
+                "monoandroid" => HandleAndroid(framework),
+                "xamarin.ios" => HandleiOS(),
+                "xamarin.tvos" => HandleTVOS(),
+                "xamarin.watchos" => HandleWatchOS(),
+                "xamarin.mac" => HandleMac(),
+                _ => Array.Empty<string>(),
+            };
 
             return FileSystemHelpers.GetSubdirectoriesWithMatch(folders, AssemblyHelpers.AssemblyFileExtensionsSet);
         }
@@ -61,7 +51,6 @@ namespace ReactiveMarbles.NuGet.Helpers
             return new[] { Path.Combine(referenceAssembliesLocation, "Xamarin.TVOS") };
         }
 
-        [SuppressMessage("StyleCop.CSharp.NamingRules", "SA1300:Element should begin with upper-case letter", Justification = "iOS special naming scheme.")]
         private static IEnumerable<string> HandleiOS()
         {
             var referenceAssembliesLocation = GetReferenceLocation("/Library/Frameworks/Xamarin.iOS.framework/Versions/Current/lib/mono/");
@@ -69,7 +58,6 @@ namespace ReactiveMarbles.NuGet.Helpers
             return new[] { Path.Combine(referenceAssembliesLocation, "Xamarin.iOS") };
         }
 
-        [SuppressMessage("StyleCop.CSharp.NamingRules", "SA1300:Element should begin with upper-case letter", Justification = "iOS special naming scheme.")]
         private static IEnumerable<string> HandleMac()
         {
             var referenceAssembliesLocation = GetReferenceLocation("/Library/Frameworks/Xamarin.Mac.framework/Versions/Current/lib/mono/");
